[FALP-40] Changing ApplicationLinksSection background color, adding padding
Replaced old getBackground function (which previously took in parameters grey and darkgrey) and replaced with new getBackground function taking in lightblue. Updated index.jsx to call new function instead (of passing down parameter with grey). Added padding to both the top and bottom of the section, so text does not stick to the top. Thought process was relatively straightforward, as much of the styling was already there (just changing to fit new specifications).

Testing: Tried different screen sizings to ensure that section was still readable and easy to understand.